requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
python-dotenv
faiss-cpu
